Ta bort poster, om det är en 'parent' så tas child post också bort

+++ quizsidan/drop_table.php
@@ -5,13 +5,15 @@ include('check_login.php');
 
 // drop table
 try {
-    $sql = "DROP TABLE " . $_POST['table'];
-    $stmt = $dbconn->prepare($sql);
-    $stmt->execute();
-
-    // create table
-    include('quiz_table.php');
+    if (isset($_POST['droptable']) && $_POST['droptable'] == 'true') {
+        $sql = "DROP TABLE " . $_POST['table'];
+        $stmt = $dbconn->prepare($sql);
+        $stmt->execute();
+        // create table
+        include('quiz_table.php');
+    }else{
+        header("Location: index.php");
+    }
 } catch (PDOException $e) {
     echo $e->getMessage();
 }
-
